First steps towards a C version of Moonlime
It occurred to me a few days ago that

 * C code would probably be faster than Lua code, and would have one less
   external dependency
 * The current parser code is pretty complex, it's not always very readable,
   and changing it can be a bit of a chore.

So, I'm rewriting Moonlime in C. Now if only there was a program that took a
high-level description of a file's structure and automated the tokenizer-
generation... Oh, wait, I've just *written* one of those; it's called Moonlime
(Lua iteration). So I'm going the bootstrap route; the C version of Moonlime
will make use of the Lua version of Moonlime until it's fully self-hosting.